Add files via upload
For use with Oyster in access.pro to change or update wavelengths observed during an NPOI observation. Generally used to update results based on FTS results. Also, in case the wrong setup was used. E.g. imaging was set in the system.config file but H-alpha was actually set for the spectrometer translation.