Add files via upload
I wrote this to take a generated ephemeris file for Jupiter that, among other things, contains the geocentric planet's RA and DEC and MJD and Year, month and Day at zero hours UT. I then wanted to get the observed LST and Hour Angle at Flagstaff for Jupiter. This is easily changed for another site by editing the east longitude. 
This requires the ct2lst.pro code from the idlastro code library.  

Included with this is a ample ephermeis file for Jupiter. The user wil make an IDL templte to read this in, so any ephemeris coudl in principle be used with a few modifications.