Simpson's rule example
Example IDL code I made that uses Simpson's rule, a built in function QSIMP in IDL, to numerically integrate an analytical function. Parking it here for future reference. 
An IDL FUNCTION is used to define the function I want to integrate and my procedure calls QSIMP with this function and the limits of integration and a few QSIMP parameters set via keywords.